adding TableElement, table commands, position commands

+++ graphics-pdfbox-layout/src/main/java/org/xbib/graphics/pdfbox/layout/table/AbstractTextCell.java
@@ -15,7 +15,7 @@ public abstract class AbstractTextCell extends AbstractCell {
         return parameters.getFont();
     }
 
-    public Integer getFontSize() {
+    public Float getFontSize() {
         return parameters.getFontSize();
     }
